pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tsconformal"
version = "0.1.0"
description = "Segmented Conformal Transport for predictive CDF recalibration under nonstationarity"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "typing_extensions>=4.5",
]

[project.optional-dependencies]
plots = ["matplotlib>=3.7", "statsmodels>=0.14"]
detectors = ["ruptures>=1.1"]
fm = [
    "chronos-forecasting>=2.0",
    "torch>=2.0",
    "transformers>=4.30",
]
dev = [
    "pytest>=7.4",
    "hypothesis>=6.80",
    "pytest-cov>=4.1",
    "mypy>=1.5",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
